+++ src/__init__.py
@@ -1,4 +1,4 @@
 """Fexobooth - Photobooth Software für fexobox"""
 
-__version__ = "2.3.1"
+__version__ = "2.3.2"
 __author__ = "fexon e.K."